renaming sentinel1_isce2.py to s1_isce2.py
